Make the loader able to load nested modules. Disable explicit importing of __init__.py when populating the loader.
Restructured class munin to more correctly not repopulate on every reboot. It may be necessary to change this in the future. Also refactored database cursor creation into its own method.

Made the loader capable of fetching all submodules to a given name, so that command can fetch munin.mod.* from it, instead of handling this itself.