Sleep for dynamic lengths of time when ticks are stale
Instead of always waiting one minute, we should try to aim for the
"next time we expect a tick to happen". We modulate this by 15 minutes
to ensure we don't get stuck fetching ticks very late due to a
temporary malfunction.

We also bail completely if we haven't found a new tick in 6 hours or
more. At this point we should probably be sending a notification
because either PA has gone down or more likely the round has ended.