New command !hugin
Adds a new command, hugin, which interacts with crontab to manage the
dumpfile loader. Assumes that run_hugin.sh is the first command in the
running user's crontab. Allows you to disable (stop), enable (start)
or view the status of the run_hugin.sh crontab.

Future work may include making the command detect the correct
placement of the run_hugin.sh command in the user's crontab.