!thinkforme: Remove 'average(ish)' from help text
I wasn't sure 'expected average roids' was right before (hence the 'ish'), but
after thinking on it some more I'm now sure that even 'ish' is an overpromise.
If a planet's roids fluctuate, then the number of guards it needs to reach a
certain alert changes too, and that change messes up all kinds of stuff in the
calculations. So I changed it to 'expected roids': a flat number that the
calculations assume you have for the remainder of the round. That's not
realistic, but at least it's accurate. Are you still reading? How are you doing?